vcpn and shared
