+++ modules/040-control-plane-manager/images/control-plane-manager/src/internal/operations/dedup.go
@@ -1,0 +1,62 @@
+/*
+Copyright 2026 Flant JSC
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package operations
+
+import (
+	controlplanev1alpha1 "control-plane-manager/api/v1alpha1"
+)
+
+func Covered(current []controlplanev1alpha1.ControlPlaneOperation, d Decision) bool {
+	return hasActiveOperation(current, d.component, d.match)
+}
+
+func (d Decision) match(op *controlplanev1alpha1.ControlPlaneOperation) bool {
+	switch d.kind {
+	case KindObserve:
+		return true
+	case KindCertRenew:
+		return hasRenewalStep(op)
+	case KindSignatureRenew:
+		return hasSignatureStep(op)
+	default:
+		return op.Spec.DesiredConfigChecksum == d.intended.Config &&
+			op.Spec.DesiredPKIChecksum == d.intended.PKI &&
+			op.Spec.DesiredCAChecksum == d.intended.CA
+	}
+}
+
+func hasRenewalStep(op *controlplanev1alpha1.ControlPlaneOperation) bool {
+	return op.HasStep(controlplanev1alpha1.StepRenewPKICerts) ||
+		op.HasStep(controlplanev1alpha1.StepRenewKubeconfigs)
+}
+
+func hasSignatureStep(op *controlplanev1alpha1.ControlPlaneOperation) bool {
+	return op.HasStep(controlplanev1alpha1.StepRenewSignature)
+}
+
+func hasActiveOperation(current []controlplanev1alpha1.ControlPlaneOperation, component controlplanev1alpha1.OperationComponent, match func(*controlplanev1alpha1.ControlPlaneOperation) bool) bool {
+	for i := range current {
+		op := &current[i]
+		if op.IsTerminal() || op.Spec.Component != component {
+			continue
+		}
+		if match(op) {
+			return true
+		}
+	}
+	return false
+}
